+++ lib/core/facets/dir/recurse.rb
@@ -1,35 +1,17 @@
+require 'facets/dir/find'
+
 class Dir
 
-  # Recursively scan a directory and pass each file to the given block.
-  #
-  #   Dir.recurse('.') do |path|
-  #     # ...
-  #   end
-  #
-  # CREDIT: George Moschovitis
-  #
-  # TODO: If fully compatible, reimplement as alias of Find.find,
-  # or just copy and paste Find.find code here if it looks more robust.
-  #
+  # @deprecated Use Dir.find or Find.find instead.
   def self.recurse(path='.', &block)
-    list = []
-    stoplist = ['.', '..']
-    Dir.foreach(path) do |f|
-      next if stoplist.include?(f)
-      filename = (path == '.' ? f : path + '/' + f)
-      list << filename
-      block.call(filename) if block
-      if FileTest.directory?(filename) and not FileTest.symlink?(filename)
-        list.concat(Dir.recurse(filename, &block))
-      end
-    end
-    list
+    warn "Dir.recurse is deprecated. Use Dir.find or Find.find instead.", uplevel: 1
+    Dir.find(path, &block)
   end
 
-  # Same as Dir#recurse.
+  # @deprecated Use Dir.find or Find.find instead.
   def self.ls_r(path='.', &block)
-    recurse(path, &block)
+    warn "Dir.ls_r is deprecated. Use Dir.find or Find.find instead.", uplevel: 1
+    Dir.find(path, &block)
   end
 
 end
-